# openapi-eventia.yaml
openapi: 3.0.1
info:
  title: Eventia API
  description: API REST para la gestión de eventos, artistas, recintos, usuarios y reservas.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /artistas:
    get:
      tags: [Artistas]
      summary: Obtener todos los artistas
      description: Filtra opcionalmente por nombreArtistico, generoMusical y activo.
      parameters:
        - name: nombreArtistico
          in: query
          required: false
          description: Filtra por nombre artístico.
          schema:
            type: string
          example: Aitana
        - name: generoMusical
          in: query
          required: false
          description: Filtra por género musical.
          schema:
            type: string
          example: Pop
        - name: activo
          in: query
          required: false
          description: Filtra por si el artista está activo.
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de artistas obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artista'
              examples:
                sinFiltros:
                  summary: Consulta sin filtros
                  value:
                    - id: 1
                      nombreArtistico: Aitana
                      nombreReal: Aitana Ocaña Morales
                      generoMusical: Pop
                      fechaNacimiento: '1999-06-27'
                      activo: true
                      cache: 5000.0
                      eventosRealizados: 20
                    - id: 2
                      nombreArtistico: Bad Bunny
                      nombreReal: Benito Antonio Martinez
                      generoMusical: Reggaeton
                      fechaNacimiento: '1994-03-10'
                      activo: true
                      cache: 15000.0
                      eventosRealizados: 50
                filtradoPorGenero:
                  summary: Consulta filtrada por género Pop
                  value:
                    - id: 1
                      nombreArtistico: Aitana
                      nombreReal: Aitana Ocaña Morales
                      generoMusical: Pop
                      fechaNacimiento: '1999-06-27'
                      activo: true
                      cache: 5000.0
                      eventosRealizados: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Artistas]
      summary: Crear artista
      description: Registra un nuevo artista.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artista'
            examples:
              ejemplo1:
                summary: Artista pop
                value:
                  nombreArtistico: Aitana
                  nombreReal: Aitana Ocaña Morales
                  generoMusical: Pop
                  fechaNacimiento: '1999-06-27'
                  activo: true
                  cache: 5000.0
                  eventosRealizados: 20
              ejemplo2:
                summary: Artista reggaeton
                value:
                  nombreArtistico: Bad Bunny
                  nombreReal: Benito Antonio Martinez
                  generoMusical: Reggaeton
                  fechaNacimiento: '1994-03-10'
                  activo: true
                  cache: 15000.0
                  eventosRealizados: 50
      responses:
        '201':
          description: Artista creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
              examples:
                ejemplo1:
                  summary: Artista pop creado
                  value:
                    id: 1
                    nombreArtistico: Aitana
                    nombreReal: Aitana Ocaña Morales
                    generoMusical: Pop
                    fechaNacimiento: '1999-06-27'
                    activo: true
                    cache: 5000.0
                    eventosRealizados: 20
                ejemplo2:
                  summary: Artista reggaeton creado
                  value:
                    id: 2
                    nombreArtistico: Bad Bunny
                    nombreReal: Benito Antonio Martinez
                    generoMusical: Reggaeton
                    fechaNacimiento: '1994-03-10'
                    activo: true
                    cache: 15000.0
                    eventosRealizados: 50
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /artistas/{id}:
    get:
      tags: [Artistas]
      summary: Obtener artista por id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Artista obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
              examples:
                ejemplo1:
                  summary: Artista pop
                  value:
                    id: 1
                    nombreArtistico: Aitana
                    nombreReal: Aitana Ocaña Morales
                    generoMusical: Pop
                    fechaNacimiento: '1999-06-27'
                    activo: true
                    cache: 5000.0
                    eventosRealizados: 20
                ejemplo2:
                  summary: Artista reggaeton
                  value:
                    id: 2
                    nombreArtistico: Bad Bunny
                    nombreReal: Benito Antonio Martinez
                    generoMusical: Reggaeton
                    fechaNacimiento: '1994-03-10'
                    activo: true
                    cache: 15000.0
                    eventosRealizados: 50
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Artistas]
      summary: Modificar artista
      description: Sustituye completamente los datos de un artista existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artista'
            examples:
              ejemplo1:
                summary: Actualizar género y caché
                value:
                  nombreArtistico: Aitana
                  nombreReal: Aitana Ocaña Morales
                  generoMusical: Pop latino
                  fechaNacimiento: '1999-06-27'
                  activo: true
                  cache: 8000.0
                  eventosRealizados: 25
              ejemplo2:
                summary: Desactivar artista
                value:
                  nombreArtistico: Bad Bunny
                  nombreReal: Benito Antonio Martinez
                  generoMusical: Latin Trap
                  fechaNacimiento: '1994-03-10'
                  activo: false
                  cache: 15000.0
                  eventosRealizados: 50
      responses:
        '200':
          description: Artista modificado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
              examples:
                ejemplo1:
                  summary: Artista actualizado
                  value:
                    id: 1
                    nombreArtistico: Aitana
                    nombreReal: Aitana Ocaña Morales
                    generoMusical: Pop latino
                    fechaNacimiento: '1999-06-27'
                    activo: true
                    cache: 8000.0
                    eventosRealizados: 25
                ejemplo2:
                  summary: Artista desactivado
                  value:
                    id: 2
                    nombreArtistico: Bad Bunny
                    nombreReal: Benito Antonio Martinez
                    generoMusical: Latin Trap
                    fechaNacimiento: '1994-03-10'
                    activo: false
                    cache: 15000.0
                    eventosRealizados: 50
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Artistas]
      summary: Modificar parcialmente un artista
      description: Actualiza uno o varios campos de un artista existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              ejemplo1:
                summary: Cambiar género musical
                value:
                  generoMusical: Pop latino
              ejemplo2:
                summary: Desactivar artista
                value:
                  activo: false
      responses:
        '200':
          description: Artista actualizado parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artista'
              examples:
                ejemplo1:
                  summary: Género actualizado
                  value:
                    id: 1
                    nombreArtistico: Aitana
                    nombreReal: Aitana Ocaña Morales
                    generoMusical: Pop latino
                    fechaNacimiento: '1999-06-27'
                    activo: true
                    cache: 5000.0
                    eventosRealizados: 20
                ejemplo2:
                  summary: Artista desactivado
                  value:
                    id: 2
                    nombreArtistico: Bad Bunny
                    nombreReal: Benito Antonio Martinez
                    generoMusical: Reggaeton
                    fechaNacimiento: '1994-03-10'
                    activo: false
                    cache: 15000.0
                    eventosRealizados: 50
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Artistas]
      summary: Eliminar artista
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Artista eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /artistas/activos:
    get:
      tags: [Artistas]
      summary: Obtener artistas activos
      description: Devuelve los artistas con activo = true mediante una consulta JPQL.
      responses:
        '200':
          description: Lista de artistas activos obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artista'
              examples:
                ejemplo1:
                  summary: Un artista activo
                  value:
                    - id: 1
                      nombreArtistico: Aitana
                      nombreReal: Aitana Ocaña Morales
                      generoMusical: Pop
                      fechaNacimiento: '1999-06-27'
                      activo: true
                      cache: 5000.0
                      eventosRealizados: 20
                ejemplo2:
                  summary: Varios artistas activos
                  value:
                    - id: 1
                      nombreArtistico: Aitana
                      nombreReal: Aitana Ocaña Morales
                      generoMusical: Pop
                      fechaNacimiento: '1999-06-27'
                      activo: true
                      cache: 5000.0
                      eventosRealizados: 20
                    - id: 3
                      nombreArtistico: Rosalia
                      nombreReal: Rosalía Vila Tobella
                      generoMusical: Flamenco Pop
                      fechaNacimiento: '1992-09-25'
                      activo: true
                      cache: 10000.0
                      eventosRealizados: 35
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Identificador único del recurso
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    BadRequest:
      description: Error de validación o petición incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            title: bad-request
            message: Error de validación
            errors:
              email: El email no es válido
              password: La contraseña debe tener entre 6 y 20 caracteres
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            title: not-found
            message: Recurso no encontrado
            errors: {}
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            title: internal-server-error
            message: Error interno del servidor
            errors: {}
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        title:
          type: string
          example: bad-request
        message:
          type: string
          example: Error de validación
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - code
        - title
        - message
        - errors
...
    Artista:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombreArtistico:
          type: string
          minLength: 2
          example: Aitana
        nombreReal:
          type: string
          minLength: 2
          example: Aitana Ocaña Morales
        generoMusical:
          type: string
          example: Pop
        fechaNacimiento:
          type: string
          format: date
          example: '1999-06-27'
        activo:
          type: boolean
          example: true
        cache:
          type: number
          format: float
          minimum: 0
          example: 5000.0
        eventosRealizados:
          type: integer
          minimum: 0
          example: 20
      required:
        - nombreArtistico
        - nombreReal
...